Fix clang build
